Switch cl-el comms to RpcChannel
Instead of using a socket (and having to open a fake connection between
EL and CL), this pr switches to
https://github.com/status-im/nim-json-rpc/pull/254 for internal
communication.

Eventually, one could make this more efficient by skipping the JSON
step, but like this, we at least no longer have to open an Engine API
port which makes this setup more secure and easy to deploy (fewer open
ports).

There's also fewer potential errors to contend with and payloads don't
have to travel across the OS buffers and instead stay internal to the
process.